+++ jgclark.Dashboard/src/dataGeneration.js
@@ -8,19 +8,12 @@ import pluginJson from '../plugin.json'
 import moment from 'moment/min/moment-with-locales'
 import fm from 'front-matter' // For reviewList functionality
 import { getSettings } from './dashboardHelpers'
-import {
-  toLocaleDateString,
-  getDateStringFromCalendarFilename,
-} from '@helpers/dateTime'
+import { toLocaleDateString, getDateStringFromCalendarFilename } from '@helpers/dateTime'
 import { clo, logDebug, logError, logInfo } from '@helpers/dev'
 import { getFolderFromFilename } from '@helpers/folders'
 import { displayTitle } from '@helpers/general'
 import { getReferencedParagraphs } from '@helpers/NPnote'
-import {
-  getTasksByType, sortListBy,
-  type GroupedTasks,
-  type SortableParagraphSubset
-} from '@helpers/sorting'
+import { getTasksByType, sortListBy, type GroupedTasks, type SortableParagraphSubset } from '@helpers/sorting'
 
 //-----------------------------------------------------------------
 // Data types
@@ -69,23 +62,31 @@ export async function getDataForDashboard(): Promise<[Array<SectionDetails>, Arr
       const thisFilename = currentDailyNote.filename
       logDebug('getDataForDashboard', `Processing ${thisFilename} which has ${String(currentDailyNote.paragraphs.length)} paras`)
       // Need to filter out non-task types for following function
-      const allParas: GroupedTasks = getTasksByType(currentDailyNote.paragraphs.filter((p) => ["open", "checklist"].includes(p.type)), true)
-      const openParas: Array<SortableParagraphSubset> = (allParas.open).concat(allParas.checklist)
+      const allParas: GroupedTasks = getTasksByType(
+        currentDailyNote.paragraphs.filter((p) => ['open', 'checklist'].includes(p.type)),
+        true,
+      )
+      const openParas: Array<SortableParagraphSubset> = allParas.open.concat(allParas.checklist)
       // clo(openParas, `${(String(openParas.length))} openParas:`)
       // sort the list only by priority, otherwise leaving order the same
       const sortedOpenParas = sortListBy(openParas, ['-priority'])
       // clo(sortedOpenParas, `${(String(sortedOpenParas.length))} sortedOpenParas:`)
       sortedOpenParas.map((p) => sectionItems.push({ ID: sectionCount, content: p.content, filename: thisFilename, type: p.calculatedType }))
-      sections.push({ ID: sectionCount, name: 'Today', description: `${openParas.length} from daily note for ${toLocaleDateString(today)}`, FAIconClass: "fa-regular fa-calendar-star", sectionTitleClass: "sidebarDaily" })
+      sections.push({
+        ID: sectionCount,
+        name: 'Today',
+        description: `${openParas.length} from daily note for ${toLocaleDateString(today)}`,
+        FAIconClass: 'fa-regular fa-calendar-star',
+        sectionTitleClass: 'sidebarDaily',
+      })
       // sections.push({ ID: sectionCount, name: 'Today', description: `${openParas.length} from daily note for ${toLocaleDateString(today)}`, FAIconClass: "fa-regular fa-calendar-star", FAIconColor: "#d0703c" })
       sectionCount++
 
       // TODO: Include context for sub-tasks/checklist
       // config.includeTaskContext
 
-
       // Get count of tasks/checklists done today
-      doneCount += currentDailyNote.paragraphs.filter((p) => ["done", "checklistDone"].includes(p.type)).length
+      doneCount += currentDailyNote.paragraphs.filter((p) => ['done', 'checklistDone'].includes(p.type)).length
 
       // Get list of open tasks/checklists scheduled to today from other notes
       const refParas = currentDailyNote ? getReferencedParagraphs(currentDailyNote, false) : []
@@ -94,7 +95,13 @@ export async function getDataForDashboard(): Promise<[Array<SectionDetails>, Arr
       if (refParas) {
         // $FlowFixMe[incompatible-use]
         refParas.map((p) => sectionItems.push({ ID: sectionCount, content: p.content, filename: p.note.filename, type: p.type }))
-        sections.push({ ID: sectionCount, name: 'Today', description: `${refParas.length} scheduled to today from other notes`, FAIconClass: "fa-regular fa-clock", sectionTitleClass: "sidebarDaily" })
+        sections.push({
+          ID: sectionCount,
+          name: 'Today',
+          description: `${refParas.length} scheduled to today from other notes`,
+          FAIconClass: 'fa-regular fa-clock',
+          sectionTitleClass: 'sidebarDaily',
+        })
         // sections.push({ ID: sectionCount, name: 'Today', description: `${refParas.length} scheduled to today from other notes`, FAIconClass: "fa-regular fa-clock", FAIconColor: "#d0703c" })
         sectionCount++
       }
@@ -114,27 +121,42 @@ export async function getDataForDashboard(): Promise<[Array<SectionDetails>, Arr
       const dateStr = getDateStringFromCalendarFilename(currentWeeklyNote.filename)
       logDebug('getDataForDashboard', `Processing ${thisFilename} which has ${String(currentWeeklyNote.paragraphs.length)} paras`)
       // Need to filter out non-task types for following function
-      const allParas: GroupedTasks = getTasksByType(currentWeeklyNote.paragraphs.filter((p) => ["open", "checklist"].includes(p.type)), true)
-      const openParas: Array<SortableParagraphSubset> = (allParas.open).concat(allParas.checklist)
+      const allParas: GroupedTasks = getTasksByType(
+        currentWeeklyNote.paragraphs.filter((p) => ['open', 'checklist'].includes(p.type)),
+        true,
+      )
+      const openParas: Array<SortableParagraphSubset> = allParas.open.concat(allParas.checklist)
       // clo(openParas, `${(String(openParas.length))} openParas:`)
       // sort the list only by priority, otherwise leaving order the same
       const sortedOpenParas = sortListBy(openParas, ['-priority'])
       // clo(sortedOpenParas, `${(String(sortedOpenParas.length))} sortedOpenParas:`)
-      sortedOpenParas.map((p) => sectionItems.push({ ID: sectionCount, content: p.content, filename: thisFilename, type: p.calculatedType }))
-      sections.push({ ID: sectionCount, name: 'This Week', description: `${openParas.length} from weekly note ${dateStr}`, FAIconClass: "fa-regular fa-calendar-week", sectionTitleClass: "sidebarWeekly" })
+      sortedOpenParas.map((p) => sectionItems.push({ ID: sectionCount, content: p.content, filename: thisFilename, type: p.calculatedType, lineIndex: p.lineIndex }))
+      sections.push({
+        ID: sectionCount,
+        name: 'This Week',
+        description: `${openParas.length} from weekly note ${dateStr}`,
+        FAIconClass: 'fa-regular fa-calendar-week',
+        sectionTitleClass: 'sidebarWeekly',
+      })
       // sections.push({ ID: sectionCount, name: 'This Week', description: `${openParas.length} from weekly note ${dateStr}`, FAIconClass: "fa-regular fa-calendar-week", FAIconColor: "#be23b6" })
       sectionCount++
 
       // Get completed count too
-      doneCount += currentWeeklyNote.paragraphs.filter((p) => ["done", "checklistDone"].includes(p.type)).length
+      doneCount += currentWeeklyNote.paragraphs.filter((p) => ['done', 'checklistDone'].includes(p.type)).length
 
       // Get list of open tasks/checklists scheduled to this week from other notes
       // FIXME: apply logic above here
       const refParas = currentWeeklyNote ? getReferencedParagraphs(currentWeeklyNote, false) : []
       if (refParas) {
         // $FlowFixMe[incompatible-use]
-        refParas.map((p) => sectionItems.push({ ID: sectionCount, content: p.content, filename: p.note.filename, type: p.type }))
-        sections.push({ ID: sectionCount, name: 'This week', description: `${refParas.length} scheduled to this week from other notes`, FAIconClass: "fa-regular fa-clock", sectionTitleClass: "sidebarWeekly" })
+        refParas.map((p) => sectionItems.push({ ID: sectionCount, content: p.content, filename: p.note.filename, type: p.type, lineIndex: p.lineIndex }))
+        sections.push({
+          ID: sectionCount,
+          name: 'This week',
+          description: `${refParas.length} scheduled to this week from other notes`,
+          FAIconClass: 'fa-regular fa-clock',
+          sectionTitleClass: 'sidebarWeekly',
+        })
         // sections.push({ ID: sectionCount, name: 'This week', description: `${refParas.length} scheduled to this week from other notes`, FAIconClass: "fa-regular fa-clock", FAIconColor: "#be23b6" })
         sectionCount++
       }
@@ -151,7 +173,13 @@ export async function getDataForDashboard(): Promise<[Array<SectionDetails>, Arr
           // sectionItems.push({ ID: sectionCount, content: '', filename: 'CCC Areas/Services.md', type: 'note' }) // Example for testing only
           sectionItems.push({ ID: sectionCount, content: '', filename: n.filename, type: 'review-note' })
         }
-        sections.push({ ID: sectionCount, name: 'Projects', description: `Next ${nextNotesToReview.length} projects ready to review`, FAIconClass: "fa-regular fa-calendar-check", sectionTitleClass: "sidebarYearly" }) // or "fa-solid fa-calendar-arrow-down" ?
+        sections.push({
+          ID: sectionCount,
+          name: 'Projects',
+          description: `Next ${nextNotesToReview.length} projects ready to review`,
+          FAIconClass: 'fa-regular fa-calendar-check',
+          sectionTitleClass: 'sidebarYearly',
+        }) // or "fa-solid fa-calendar-arrow-down" ?
         // sections.push({ ID: sectionCount, name: 'Projects', description: `Next ${nextNotesToReview.length} projects ready to review`, FAIconClass: "fa-regular fa-calendar-check", FAIconColor: "#bc782e" }) // or "fa-solid fa-calendar-arrow-down" ?
         sectionCount++
       } else {
@@ -160,13 +188,12 @@ export async function getDataForDashboard(): Promise<[Array<SectionDetails>, Arr
     }
 
     // Let's also pass count of completed things, to help morale!
-    sections.push({ ID: sectionCount, name: 'Done', description: String(doneCount), FAIconClass: "fa-regular fa-circle-check", sectionTitleClass: "" })
+    sections.push({ ID: sectionCount, name: 'Done', description: String(doneCount), FAIconClass: 'fa-regular fa-circle-check', sectionTitleClass: '' })
     sectionCount++
 
     logDebug('getDataForDashboard', `getDataForDashboard finished, with ${String(sections.length)} sections and ${String(sectionItems.length)} items`)
     return [sections, sectionItems]
-  }
-  catch (error) {
+  } catch (error) {
     logError(pluginJson, error.message)
     return [[], []] // for completeness
   }
@@ -189,14 +216,13 @@ export async function logDashboardData(): Promise<void> {
         console.log(`- ${item?.content ?? ''} ${item.filename}`)
       }
     }
-  }
-  catch (error) {
+  } catch (error) {
     logError('pluginJson', error.message)
   }
 }
 
 //-------------------------------------------------------------------------------
-/** 
+/**
  * Get list of the next note(s) to review (if any).
  * It assumes the full-review-list exists and is sorted by nextReviewDate (earliest to latest).
  * Note: This is a variant of the original singular version in jgclark.Reviews/src/reviews.js
@@ -247,8 +273,7 @@ async function getNextNotesToReview(numToReturn: number): Promise<Array<TNote>> 
         return []
       }
     }
-  }
-  catch (error) {
+  } catch (error) {
     logError(pluginJson, `dashboard::getNextNotesToReview: ${error.message}`)
     return []
   }
@@ -264,4 +289,4 @@ function stripMailtoLinks(email: string): string {
 // Convert markdown links to HTML links
 function convertMarkdownLinks(text: string): string {
   return text.replace(/\[([^\]]+)\]\(([^\)]+)\)/g, '<a href="$1">$2</a>')
-} 
+}