+++ np.Shared/src/index.js
@@ -8,17 +8,15 @@
 
 const pluginID = 'np.Shared'
 import pluginJson from '../plugin.json'
-import { getPluginJson } from '@helpers/NPConfiguration'
+import { getPluginJson, updateSettingData } from '@helpers/NPConfiguration'
 import { clo, JSP, logDebug, logError, logInfo } from '@helpers/dev'
-import { updateSettingData } from '@helpers/NPConfiguration'
 import { showMessage } from '@helpers/userInput'
 
 export function fakeInstall(): void {
   try {
     logDebug(pluginID, 'fakeInstall: Starting ...')
     // TODO: stuff
-  }
-  catch (error) {
+  } catch (error) {
     logError(pluginID, JSP(error))
   }
 }
@@ -32,8 +30,7 @@ export async function logProvidedResources(): Promise<void> {
     const livePluginJson = await getPluginJson(pluginID)
     const requiredFiles = livePluginJson['plugin.requiredFiles']
     logInfo(pluginID, `Resources Provided by np.Shared according to its plugin.json file:\n- ${requiredFiles.join('\n- ')}`)
-  }
-  catch (error) {
+  } catch (error) {
     logError(pluginID, JSP(error))
   }
 }
@@ -42,8 +39,7 @@ export function logLocallyAvailableResources(): void {
   try {
     const availableFiles = ['test', 'me'] // FIXME:
     logInfo(pluginID, `# Locally available resources, as found in Plugins/np.Shared/:\n- ${availableFiles.join('\n- ')}`)
-  }
-  catch (error) {
+  } catch (error) {
     logError(pluginID, JSP(error))
   }
 }
@@ -66,9 +62,7 @@ export async function onUpdateOrInstall(): Promise<void> {
 
     // Tell user the plugin has been updated
     if (pluginJson['plugin.lastUpdateInfo'] !== 'undefined') {
-      await showMessage(pluginJson['plugin.lastUpdateInfo'], 'OK, thanks',
-        `Plugin ${pluginJson['plugin.name']}\nupdated to v${pluginJson['plugin.version']}`
-      )
+      await showMessage(pluginJson['plugin.lastUpdateInfo'], 'OK, thanks', `Plugin ${pluginJson['plugin.name']}\nupdated to v${pluginJson['plugin.version']}`)
     }
   } catch (error) {
     logError(pluginID, error)