+++ np.Shared/CHANGELOG.md
@@ -2,8 +2,17 @@
 
 See [Shared Plugin's README](https://github.com/NotePlan/plugins/blob/main/np.Shared/README.md) for details on this plugin.
 
-## [0.1.0] - 2023-02-04 (@jgclark)
+## [0.2.0] - 2023-02-18 (@dwertheimer)
+
 ### Added
+
+- File pluginToHTMLCommsBridge.js
+
+## [0.1.0] - 2023-02-04 (@jgclark)
+
+### Added
+
 First release:
+
 - provides the **FontAwesome** assets
 - **/logShared** command: lists all shared assets in the Plugin Console.